refactor(chatpanel): split tab-open atoms by target surface
Move the 650-line chatPanelTabOpenAtoms.ts into chatPanelTabOpen/ with one
module per surface (startPage, session, workManagement, integrations);
chatPanelTabOpenAtoms.ts stays as a re-export barrel so all 8 import sites
are unchanged. Pure code motion, no behaviour change.
